Create 2DArray
This program will manipulate arrays by filling them with random values and performing operations (sum, average median, finding the largest value, and the number of times the largest value occurs in the array)on the array. The 1D array values will also be placed into a 2D array. This program will also read and write the 2D array to and from a file.