Create arrayLists
This program will create a model of the ocean in the form of a grid of elements, either ships or land. The file specifies the grid size, reading in the number of rows and columns to create an ocean object. This grid will be filled partially with land and ships with their locations based on their file's specified location. When the ocean has been filled with ships and land the grid will be displayed and a ship report will be printed. The ship report will display ships in order from largest to smallest capacity.