Create genericClassAndJCFStack
This program will implement and use a generic class and generic methods. It will also use the JCF stack to place values onto the stack and remove odd values without changing the order of the even values and display. This program will also fill a stack with values read from the file and sort the stack by removing repeatedly the smallest element. The stack will be organized with the smallest value at the top in ascending order.