Create binarySearchTreeWithParrots
This program will fill a binary search tree with parrots. Each parrot will read from the file their ID number, name, and song when the parrots are placed in the tree they will be ordered by their IDs. After parrots are placed in the tree the program will use level-order to display the parrots' song and list parrot names using the leaves of the tree going left to right