Create studentAndClassObjects
This program works with array objects in the main and an object that contains an array. This is broken into two tasks. #1 Using Student class to create student objects and store them in an array in main. Setters and getters will be used to get/set information out of the class into main and find the highest GPA and average from students in an array. #2 using Course class to create two-course classes that hold an array of students as an instance variable in the class. The course class method will be used to add students to the array in class. This class will get the highest GPA and average GPA from the students in the course.